requirements.txt
pyrogram>=2.0.106
openai>=1.14.3
APScheduler>=3.10.4
python-dotenv>=1.0.1
pytz>=2024.1
fastapi>=0.100.0
uvicorn>=0.23.2
